+++ src/lib.rs
@@ -1,5 +1,5 @@
 #![doc(
-    html_root_url = "https://docs.rs/arc-swap/0.3.5/arc-swap/",
+    html_root_url = "https://docs.rs/arc-swap/0.3.6/arc-swap/",
     test(attr(deny(warnings)))
 )]
 #![deny(missing_docs, warnings)]